[ELASTICSEARCH] Remove manipulating the bin scripts, use ES_HOME variable
Instead of hardcoding paths to Homebrew locations in the `bin/elasticsearch`
and `bin/plugin` scripts, set up the ES_HOME environment variable
in the `bin/elasticsearch.in.sh` properly.

Also, simplify the Regex for replacing path to
Elasticsearch classes (`--classpath`) in the `bin/plugin` script.

NOTE: We still need to manipulate the `bin/plugin` script, because
      it does not load the ES_INCLUDE file.